pyproject.toml
[tool.poetry]
name = "mlops-course-python-best-practice"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "src" }] 

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.3.0"
torchvision = "^0.18.0"
pillow = "^10.3.0"
mypy = "^1.10.0"
ruff = "^0.4.5"
pytest = "^8.2.1"
pytest-cov = "^5.0.0"
pre-commit = "^3.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = [
  "tests"
]